Add user agent for the notebook magic.
This overwrites the client library UA to the same string, but with
`-magic`, in order to distinguish the two.

It's done during import-time so that it works for both `%load_ext` and
`import ...magics`, and so that it still allows a user to call
`palm.configure(...)` before using the magic and provide their own
client, UA, etc.

Change-Id: I559b669495cf92aaf1c750a527b6bac9c73bd00f